+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### v2.0.1
+- Use Android Q internal code name 
 - Fix Android O
 
 ### v2.0.0 (2023-10-08)