+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added Finnish translation
 - Added Thai translation
 - Added Norwegian translation
+- Added Filipino translation
 - Fix Android O
 
 ### v2.0.0 (2023-10-08)